scripts/requirements.txt
Jinja2==3.1.4
MarkupSafe==2.1.5
PyYAML==6.0.1
typeguard==4.3.0
typing_extensions==4.12.2
